blockchain/requirements.txt
web3==6.0.0
python-dotenv==0.19.0
py-solc-x==1.1.1
